Implement application registration with Gtk.Application
This adds an optional application_id field in PEWApp. If this is set to
some well-known string in a subclass, GTK will ensure that only one
instance with that application_id is running. Applications using
pyeverywhere with GTK can set application_id to a value such as
"com.example.Application" to support this feature.