Support PEWShortcut in the GTK front-end
Some key names need to be translated to the names used by GTK, such as
"plus" and "minus". This can be done strictly inside the GTK module, on
assumption that these keys are either provided as ascii characters (in
the case of punctuation), or as commonly understood names (in the case
of F1 and Delete) that are the same in the different toolkits we
support. In the event that a key name from the application is not
understood by GTK, a warning will be displayed.